Added Date Store Observer
