+++ src/main/java/com/miniconomy/commercial_bank_service/simulation_management/observer/DateStoreObserver.java
@@ -1,0 +1,7 @@
+package com.miniconomy.commercial_bank_service.simulation_management.observer;
+
+public abstract class DateStoreObserver {
+    
+    public abstract void update(String newDate);
+
+}
